Reconfigure for shared logging module
