+++ setup.cfg
@@ -21,10 +21,12 @@ packages =
 
 [entry_points]
 console_scripts =
-    discord-fate-bot = discord_fate_bot:main
-    discord-fate-bot-migrate = discord_fate_bot.migrations:main
+    discord-fate-bot = discord_fate_bot.console_script:main
+    discord-fate-bot-migrate = discord_fate_bot.migrations.console_script:main
 
 [pbr]
 skip_authors = True
 skip_changelog = True
+
+# vim: set ft=dosini:
